scripts/requirements.txt
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
torch>=2.1.0
pillow>=10.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
